backporting dry run and checks

+++ Code.gs
@@ -1,12 +1,16 @@
-// Modify the calendar name and search query as desired
+// Delete 加油 ("jiā yóu") Events
 // WARNING: Deleted events are not recoverable! Once they are deleted, they are gone forever!
 
+// Search parameters
 var myCalendarName = "JIA YOU"; // Must name it differently from the owner name
-var myNewQuery = "Updated Meeting"; // Query ignores any extra spacing
-var myNewQueryAdd = "J Day"; // Query ignores any extra spacing, input additional query to confine search
+var myNewQuery = "Updated Meeting"; // Event you want to delete, query ignores any extra spacing
+var myNewQueryAdd = "J Day"; // Input additional query to confine search, query ignores any extra spacing
 var myNewStart = ""; // Confine date range
 var myNewEnd = ""; // Confine date range
-// Accepted date formats: Mmm DD YYYY; MM/DD/YYYY; DD Mmm YYYY
+// Accepted date formats: Mmm DD YYYY, MM/DD/YYYY, DD Mmm YYYY
+var myNewDryRun = false; // test script before running it in production
+
+
 
 
 
@@ -16,19 +20,27 @@ var myNewEnd = ""; // Confine date range
 
 function deleteEvents() {
   var calendarName = myCalendarName;
-  var calendars = CalendarApp.getAllCalendars();  // Get all calendars
-  
+  var calendars = CalendarApp.getAllCalendars(); // Get all calendars
+  var calendarId = ""; // Initially null
+
   // Loop through all calendars and find the one with the matching name
   for (var i = 0; i < calendars.length; i++) {
     if (calendars[i].getName() === calendarName) {
-      Logger.log("Calendar ID for \"" + calendarName + "\": " + calendars[i].getId());
-      var calendarId = String(calendars[i].getId());  // Assign the calendar ID
+      // Logger.log(
+      //   'Calendar ID for "' + calendarName + '": ' + calendars[i].getId()
+      // );
+      calendarId = String(calendars[i].getId()); // Assign the calendar ID
     }
+  }
+
+  // Check if loop finds no calendar
+  if (calendarId === "") {
+    return "No \"" + calendarName + "\" calendar exists!";
   }
 
   // Access the calendar
   var calendar = CalendarApp.getCalendarById(calendarId);
-  
+
   // Check for null dates
   if (myNewStart !== "" && myNewEnd !== "") {
     // Set the search parameters
@@ -39,11 +51,11 @@ function deleteEvents() {
     myNewEnd.setDate(myNewEnd.getDate() + 1); // include end date in search
 
     // Search for events between start and end dates
-    var events = calendar.getEvents(myNewStart, myNewEnd, {search: query});
+    var events = calendar.getEvents(myNewStart, myNewEnd, { search: query });
 
     // Check additional query
     if (queryAdd !== "") {
-      var eventsAdd = calendar.getEvents(myNewStart, myNewEnd, {search: queryAdd});
+      var eventsAdd = calendar.getEvents(myNewStart, myNewEnd, { search: queryAdd });
     }
   } else {
     // Set the search parameters
@@ -54,14 +66,28 @@ function deleteEvents() {
     oneYearFromNow.setFullYear(now.getFullYear() + 1);
 
     // Search for events between now and one year from now
-    var events = calendar.getEvents(now, oneYearFromNow, {search: query});
+    var events = calendar.getEvents(now, oneYearFromNow, { search: query });
 
     // Check additional query
     if (queryAdd !== "") {
-      var eventsAdd = calendar.getEvents(now, oneYearFromNow, {search: queryAdd});
+      var eventsAdd = calendar.getEvents(now, oneYearFromNow, {
+        search: queryAdd,
+      });
     }
   }
-  
+
+  // Check if query finds no events
+  if (events.length === 0) {
+    return "No \"" + query + "\" events exist!";
+  }
+  // Check if queryAdd finds no events
+  if (queryAdd !== "" && eventsAdd.length === 0) {
+    return "No \"" + queryAdd + "\" events exist!";
+  }
+
+  // Check if query and queryAdd find no matching events below
+  var match = "no";
+
   if (queryAdd !== "") {
     // Loop through each event found
     events.forEach(function(event) {
@@ -75,18 +101,28 @@ function deleteEvents() {
         var eventDateAdd = eventAdd.getStartTime();
 
         // Extract just the date part as a string, again
-        var eventDateAdd = eventDateAdd.toDateString();
+        eventDateAdd = eventDateAdd.toDateString();
 
         // Find matches
         if (eventDate === eventDateAdd) {
           // Delete the event
-          event.deleteEvent(); // Gone forever!
+          if (!myNewDryRun) {
+            event.deleteEvent(); // Gone forever!
+          }
+
+          match = "yes";
 
           // Log which events were deleted
           Logger.log("Deleted an event on " + eventDate + ".");
         }
       });
     });
+    if (match === "no") {
+      return "No \"" + query + "\" and \"" + queryAdd + "\" events match!";
+    }
+    else {
+      return "Events deleted!";
+    }
   } else {
     // Loop through each event found
     events.forEach(function(event) {
@@ -96,11 +132,13 @@ function deleteEvents() {
       eventDate = eventDate.toDateString();
 
       // Delete the event
-      event.deleteEvent(); // Gone forever!
+      if (!myNewDryRun) {
+        event.deleteEvent(); // Gone forever!
+      }
 
       // Log which events were deleted
       Logger.log("Deleted an event on " + eventDate + ".");
     });
+    return "Events deleted!";
   }
-  return "Events deleted!";
 }
